feat: Refactor purchases add page, dashboard page, and sale table for improved performance and functionality
This commit refactors the purchases add page, dashboard page, and sale table to improve page load performance and add new functionality. Specifically, lazy loading for images is implemented in the purchases add page, the dashboard page is updated to import the DashboardStats component, and the sale table now handles null order dates and converts the orderDate field to a string for search functionality.

Update react-countup dependency to version 6.5.3

This commit adds the react-countup dependency to the package.json file and updates it to version 6.5.3.

Refactor saleActions.ts to use any type for saleData parameter

In this commit, the saleActions.ts file is updated to use the any type for the saleData parameter in the updateSale function.

Update sweetalert2 dependency to version 11.6.13

The sweetalert2 dependency in package.json is updated to version 11.6.13 in this commit.

Remove unused 'use server' import statement

This commit refactors the purchases add page to remove the unused 'use server' import statement.

Update DashboardTable to handle null order dates

The SaleTable component is refactored in this commit to handle null order dates in the orderDate field.

Update orderDate rendering in SaleTable component

In this commit, the rendering of the orderDate field in the SaleTable component is updated to convert it to a string if it exists.

Update roles&status.ts and ProductionActions.ts to use explicit type annotations

The roles&status.ts and ProductionActions.ts files are refactored in this commit to use explicit type annotations for improved code clarity.

Add deliverySlice and update rootReducer to include delivery reducer

This commit adds the deliverySlice to the Redux store and updates the rootReducer to include the delivery reducer.

Update react-apexcharts dependency to version 1.4.1

The react-apexcharts dependency is added to package.json and updated to version 1.4.1 in this commit.

Update landing page banner color per client request

The landing page banner color is updated as requested by the client in this commit.
