+++ app/(Dashboard)/page.tsx
@@ -1,14 +1,10 @@
-import DashboardStats from "../components/dashboard/DashboardStats";
-import DashboardMain from "./components/DashboardMain";
+import DashboardMain from "../components/dashboard/DashboardMain";
 
 
 export default async function Home() {
 
-  // const dashboardStatistics = await fetch()
-
   return (
     <>
-      <DashboardStats />
       <DashboardMain />
     </>
   );
