+++ app/components/dashboard/AreaDonutCharts.tsx
@@ -1,0 +1,326 @@
+'use client'
+import React from 'react'
+import Dropdown from '../Layout/Dropdown'
+import dynamic from 'next/dynamic';
+import { RevenueChart, SalesByCategory } from '@/types/types';
+const ReactApexChart = dynamic(() => import('react-apexcharts'), {
+    ssr: false,
+});
+
+interface AreaDonutChartsProps {
+    isDark: boolean;
+    isMounted: boolean;
+    isRtl: boolean;
+    revenueChartData: RevenueChart | undefined;
+    salesByCategoryData: SalesByCategory | undefined;
+}
+
+
+const AreaDonutCharts = ({ isDark, isMounted, isRtl, revenueChartData, salesByCategoryData }: AreaDonutChartsProps) => {
+
+
+
+
+    //Revenue Chart
+    const revenueChart: any = {
+        series: [
+            {
+                name: 'Income',
+                data: revenueChartData?.monthlySales,
+            },
+            {
+                name: 'Expenses',
+                data: revenueChartData?.monthlyPurchases,
+            },
+        ],
+        options: {
+            chart: {
+                height: 325,
+                type: 'area',
+                fontFamily: 'Nunito, sans-serif',
+                zoom: {
+                    enabled: false,
+                },
+                toolbar: {
+                    show: false,
+                },
+            },
+
+            dataLabels: {
+                enabled: false,
+            },
+            stroke: {
+                show: true,
+                curve: 'smooth',
+                width: 2,
+                lineCap: 'square',
+            },
+            dropShadow: {
+                enabled: true,
+                opacity: 0.2,
+                blur: 10,
+                left: -7,
+                top: 22,
+            },
+            colors: isDark ? ['#2196F3', '#E7515A'] : ['#1B55E2', '#E7515A'],
+            markers: {
+                discrete: [
+                    {
+                        seriesIndex: 0,
+                        dataPointIndex: 6,
+                        fillColor: '#1B55E2',
+                        strokeColor: 'transparent',
+                        size: 7,
+                    },
+                    {
+                        seriesIndex: 1,
+                        dataPointIndex: 5,
+                        fillColor: '#E7515A',
+                        strokeColor: 'transparent',
+                        size: 7,
+                    },
+                ],
+            },
+            labels: ['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul', 'Aug', 'Sep', 'Oct', 'Nov', 'Dec'],
+            xaxis: {
+                axisBorder: {
+                    show: false,
+                },
+                axisTicks: {
+                    show: false,
+                },
+                crosshairs: {
+                    show: true,
+                },
+                labels: {
+                    offsetX: isRtl ? 2 : 0,
+                    offsetY: 5,
+                    style: {
+                        fontSize: '12px',
+                        cssClass: 'apexcharts-xaxis-title',
+                    },
+                },
+            },
+            yaxis: {
+                tickAmount: 7,
+                labels: {
+                    formatter: (value: number) => {
+                        return value / 100 + 'K';
+                    },
+                    offsetX: isRtl ? -30 : -10,
+                    offsetY: 0,
+                    style: {
+                        fontSize: '12px',
+                        cssClass: 'apexcharts-yaxis-title',
+                    },
+                },
+                opposite: isRtl ? true : false,
+            },
+            grid: {
+                borderColor: isDark ? '#191E3A' : '#E0E6ED',
+                strokeDashArray: 5,
+                xaxis: {
+                    lines: {
+                        show: false,
+                    },
+                },
+                yaxis: {
+                    lines: {
+                        show: true,
+                    },
+                },
+                padding: {
+                    top: 0,
+                    right: 0,
+                    bottom: 0,
+                    left: 0,
+                },
+            },
+            legend: {
+                position: 'top',
+                horizontalAlign: 'right',
+                fontSize: '16px',
+                markers: {
+                    width: 10,
+                    height: 10,
+                    offsetX: -2,
+                },
+                itemMargin: {
+                    horizontal: 10,
+                    vertical: 5,
+                },
+            },
+            tooltip: {
+                marker: {
+                    show: true,
+                },
+                x: {
+                    show: false,
+                },
+            },
+            fill: {
+                type: 'gradient',
+                gradient: {
+                    shadeIntensity: 1,
+                    inverseColors: !1,
+                    opacityFrom: isDark ? 0.19 : 0.28,
+                    opacityTo: 0.05,
+                    stops: isDark ? [100, 100] : [45, 100],
+                },
+            },
+        },
+    };
+
+    //Sales By Category
+
+    const salesByCategory: any = {
+        series: salesByCategoryData?.productQuantity,
+        options: {
+            chart: {
+                type: 'donut',
+                height: 460,
+                fontFamily: 'Nunito, sans-serif',
+            },
+            dataLabels: {
+                enabled: false,
+            },
+            stroke: {
+                show: true,
+                width: 25,
+                colors: isDark ? '#0e1726' : '#fff',
+            },
+            colors: isDark ? ['#5c1ac3', '#e2a03f', '#e7515a', '#e2a03f'] : ['#e2a03f', '#5c1ac3', '#e7515a'],
+            legend: {
+                position: 'bottom',
+                horizontalAlign: 'center',
+                fontSize: '14px',
+                markers: {
+                    width: 10,
+                    height: 10,
+                    offsetX: -2,
+                },
+                height: 50,
+                offsetY: 20,
+            },
+            plotOptions: {
+                pie: {
+                    donut: {
+                        size: '65%',
+                        background: 'transparent',
+                        labels: {
+                            show: true,
+                            name: {
+                                show: true,
+                                fontSize: '29px',
+                                offsetY: -10,
+                            },
+                            value: {
+                                show: true,
+                                fontSize: '26px',
+                                color: isDark ? '#bfc9d4' : undefined,
+                                offsetY: 16,
+                                formatter: (val: any) => {
+                                    return val;
+                                },
+                            },
+                            total: {
+                                show: true,
+                                label: 'Total',
+                                color: '#888ea8',
+                                fontSize: '29px',
+                                formatter: (w: any) => {
+                                    return w.globals.seriesTotals.reduce(function (a: any, b: any) {
+                                        return a + b;
+                                    }, 0);
+                                },
+                            },
+                        },
+                    },
+                },
+            },
+            labels: salesByCategoryData?.productName,
+            states: {
+                hover: {
+                    filter: {
+                        type: 'none',
+                        value: 0.15,
+                    },
+                },
+                active: {
+                    filter: {
+                        type: 'none',
+                        value: 0.15,
+                    },
+                },
+            },
+        },
+    };
+
+    return (
+        <div className="mb-6 grid gap-6 xl:grid-cols-3">
+            <div className="panel h-full xl:col-span-2">
+                <div className="mb-5 flex items-center justify-between dark:text-white-light">
+                    <h5 className="text-lg font-semibold">Revenue</h5>
+                    <div className="dropdown">
+                        <Dropdown
+                            offset={[0, 1]}
+                            placement={`${isRtl ? 'bottom-start' : 'bottom-end'}`}
+                            button={
+                                <svg className="h-5 w-5 text-black/70 hover:!text-primary dark:text-white/70" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+                                    <circle cx="5" cy="12" r="2" stroke="currentColor" strokeWidth="1.5" />
+                                    <circle opacity="0.5" cx="12" cy="12" r="2" stroke="currentColor" strokeWidth="1.5" />
+                                    <circle cx="19" cy="12" r="2" stroke="currentColor" strokeWidth="1.5" />
+                                </svg>
+                            }
+                        >
+                            <ul>
+                                <li>
+                                    <button type="button">Weekly</button>
+                                </li>
+                                <li>
+                                    <button type="button">Monthly</button>
+                                </li>
+                                <li>
+                                    <button type="button">Yearly</button>
+                                </li>
+                            </ul>
+                        </Dropdown>
+                    </div>
+                </div>
+                <p className="text-lg dark:text-white-light/90">
+                    {/* Total Profit <span className="ml-2 text-primary">$10,840</span> */}
+                </p>
+                <div className="relative">
+                    <div className="rounded-lg bg-white dark:bg-black">
+                        {isMounted ? (
+                            <ReactApexChart series={revenueChart.series} options={revenueChart.options} type="area" height={325} width={'100%'} />
+                        ) : (
+                            <div className="grid min-h-[325px] place-content-center bg-white-light/30 dark:bg-dark dark:bg-opacity-[0.08] ">
+                                <span className="inline-flex h-5 w-5 animate-spin rounded-full  border-2 border-black !border-l-transparent dark:border-white"></span>
+                            </div>
+                        )}
+                    </div>
+                </div>
+            </div>
+
+            <div className="panel h-full">
+                <div className="mb-5 flex items-center">
+                    <h5 className="text-lg font-semibold dark:text-white-light">Sales By Category</h5>
+                </div>
+                <div>
+                    <div className="rounded-lg bg-white dark:bg-black">
+                        {isMounted ? (
+                            <ReactApexChart series={salesByCategory.series} options={salesByCategory.options} type="donut" height={460} width={'100%'} />
+                        ) : (
+                            <div className="grid min-h-[325px] place-content-center bg-white-light/30 dark:bg-dark dark:bg-opacity-[0.08] ">
+                                <span className="inline-flex h-5 w-5 animate-spin rounded-full  border-2 border-black !border-l-transparent dark:border-white"></span>
+                            </div>
+                        )}
+                    </div>
+                </div>
+            </div>
+        </div>
+    )
+}
+
+export default AreaDonutCharts
